Refactor enrollment proof module packages
