+++ enrollmentproof/src/main/kotlin/com/gdavidpb/tuindice/enrollmentproof/data/repository/enrollmentproof/LocalDataSource.kt
@@ -1,4 +1,4 @@
-package com.gdavidpb.tuindice.enrollmentproof.data.enrollmentproof.source
+package com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof
 
 interface LocalDataSource {
 	suspend fun getCurrentQuarterName(uid: String): String?
