+++ enrollmentproof/src/main/kotlin/com/gdavidpb/tuindice/enrollmentproof/di/EnrollmentProofModule.kt
@@ -2,14 +2,14 @@ package com.gdavidpb.tuindice.enrollmentproof.di
 
 import com.gdavidpb.tuindice.base.BuildConfig
 import com.gdavidpb.tuindice.base.utils.extension.create
-import com.gdavidpb.tuindice.enrollmentproof.data.api.ApiDataSource
-import com.gdavidpb.tuindice.enrollmentproof.data.api.EnrollmentProofApi
-import com.gdavidpb.tuindice.enrollmentproof.data.enrollmentproof.EnrollmentProofDataRepository
-import com.gdavidpb.tuindice.enrollmentproof.data.enrollmentproof.source.LocalDataSource
-import com.gdavidpb.tuindice.enrollmentproof.data.enrollmentproof.source.RemoteDataSource
-import com.gdavidpb.tuindice.enrollmentproof.data.enrollmentproof.source.StorageDataSource
-import com.gdavidpb.tuindice.enrollmentproof.data.room.RoomDataSource
-import com.gdavidpb.tuindice.enrollmentproof.data.storage.InternalStorageDataSource
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.source.EnrollmentProofApiDataSource
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.EnrollmentProofApi
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.EnrollmentProofDataRepository
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.LocalDataSource
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.RemoteDataSource
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.StorageDataSource
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.source.RoomDataSource
+import com.gdavidpb.tuindice.enrollmentproof.data.repository.enrollmentproof.source.InternalStorageDataSource
 import com.gdavidpb.tuindice.enrollmentproof.domain.repository.EnrollmentProofRepository
 import com.gdavidpb.tuindice.enrollmentproof.domain.usecase.FetchEnrollmentProofUseCase
 import com.gdavidpb.tuindice.enrollmentproof.domain.usecase.exceptionhandler.FetchEnrollmentProofExceptionHandler
@@ -41,7 +41,7 @@ val enrollmentProofModule = module {
 	/* Data sources */
 
 	factoryOf(::RoomDataSource) { bind<LocalDataSource>() }
-	factoryOf(::ApiDataSource) { bind<RemoteDataSource>() }
+	factoryOf(::EnrollmentProofApiDataSource) { bind<RemoteDataSource>() }
 	factoryOf(::InternalStorageDataSource) { bind<StorageDataSource>() }
 
 	/* Enrollment Proof Api */
